feat($sce): handle URL sanitization through the `$sce` service
Thanks to @rjamet for the original work on this feature.

This is a large patch to handle URLs with the $sce service, similarly
to HTML context.

Where we previously sanitized URL attributes when setting attribute
value inside the
`$compile` service, we now only apply an `$sce` context requirement
and leave the
`$interpolate` service to deal with sanitization.

This commit introduces a new `$sce` context called `MEDIA_URL`, which
represents
a URL used as a source for a media element that is not expected to
execute code, such as
image, video, audio, etc.
The context hierarchy is setup so that a value trusted as `URL` is
also trusted in the
`MEDIA_URL` context, in the same way that the a value trusted as
`RESOURCE_URL` is also
trusted in the `URL` context (and transitively also the `MEDIA_URL`
context).

The `$sce` service will now automatically attempt to sanitize
non-trusted values that
require the `URL` or `MEDIA_URL` context:

* When calling `getTrustedMediaUrl()` a value that is not already a
trusted `MEDIA_URL`
will be sanitized using the `imgSrcSanitizationWhitelist`.
* When calling `getTrustedUrl()` a value that is not already a trusted
`URL` will be
sanitized using the `aHrefSanitizationWhitelist`.

This results in behaviour that closely matches the previous
sanitization behaviour.

To keep rough compatibility with existing apps, we need to allow
concatenation of values
that may contain trusted contexts. The following approach is taken for
situations that
require a `URL` or `MEDIA_URL` secure context:

* A single trusted value is trusted, e.g. `"{{trustedUrl}}"` and will
not be sanitized.
* A single non-trusted value, e.g. `"{{ 'javascript:foo' }}"`, will be
handled by
`getTrustedMediaUrl` or `getTrustedUrl)` and sanitized.
* Any concatenation of values (which may or may not be trusted)
results in a
non-trusted type that will be handled by `getTrustedMediaUrl` or
`getTrustedUrl` once the
concatenation is complete.
E.g. `"javascript:{{safeType}}"` is a concatenation of a non-trusted
and a trusted value,
which will be sanitized as a whole after unwrapping the `safeType`
value.
* An interpolation containing no expressions will still be handled by
`getTrustedMediaUrl` or
`getTrustedUrl`, whereas before this would have been short-circuited
in the `$interpolate`
service. E.g. `"some/hard/coded/url"`. This ensures that `ngHref`
and similar directives
still securely, even if the URL is hard-coded into a template or
index.html (perhaps by
server-side rendering).

BREAKING CHANGES:

If you use `attrs.$set` for URL attributes (a[href] and img[src])
there will no
longer be any automated sanitization of the value. This is in line
with other
programmatic operations, such as writing to the innerHTML of an
element.

If you are programmatically writing URL values to attributes from
untrusted
input then you must sanitize it yourself. You could write your own
sanitizer or copy
the private `$$sanitizeUri` service.

Note that values that have been passed through the `$interpolate`
service within the
`URL` or `MEDIA_URL` will have already been sanitized, so you would
not need to sanitize
these values again.