fix($templateRequest): always return the template that is stored in the cache
Previously, `$templateRequest` returned the raw `$http` response data
on the
first request for a template and then the value from the cache for
subsequent
requests.

If the value is transformed when being added to the cache (by
decorating
`$templateCache.put`)  the return value of `$templateRequest` would be
inconsistent depending upon when the request is made.

This commit ensures the cached value is returned instead of the raw
`$http`
response data, thus allowing the `$templateCache` service to be
decorated.

Closes #16225